requirements.txt
selenium
requests
2captcha-python
pillow